feat: move functions to get links to its own file
